requirements.txt
# RL Subtext Transfer Experiment Dependencies

# Core ML
torch>=2.0.0
transformers>=4.35.0

# LoRA and efficient fine-tuning
peft>=0.7.0
bitsandbytes>=0.41.0
accelerate>=0.24.0

# RL training (need 0.9+ for GRPO support)
trl>=0.9.0

# API calls
requests>=2.28.0
httpx>=0.24.0

# Config and utilities
pyyaml>=6.0
python-dotenv>=1.0.0

# Progress and logging
tqdm>=4.65.0

# Optional: Weights & Biases
# wandb>=0.15.0

# Data
numpy>=1.24.0
